Fix streams read/write + generic circuit

+++ voting-system/src/command.rs
@@ -1,3 +1,3 @@
 use druid::Selector;
 
-pub const VOTE: Selector<u8> = Selector::new("app.vote");
+pub const VOTE: Selector<u16> = Selector::new("app.vote");
